LayoutTests/svg/animations/svgpointlist-animation-2-expected.txt

SVG 1.1 dynamic animation tests

Tests from-by animation of points on polygons.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS poly.animatedPoints.getItem(2).x is 200
PASS poly.animatedPoints.getItem(2).y is 200
PASS poly.points.getItem(2).x is 200
PASS poly.points.getItem(2).y is 200
PASS poly.animatedPoints.getItem(2).x is 250
PASS poly.animatedPoints.getItem(2).y is 250
PASS poly.points.getItem(2).x is 200
PASS poly.points.getItem(2).y is 200
PASS poly.animatedPoints.getItem(2).x is 300
PASS poly.animatedPoints.getItem(2).y is 300
PASS poly.points.getItem(2).x is 200
PASS poly.points.getItem(2).y is 200
PASS poly.animatedPoints.getItem(2).x is 200
PASS poly.animatedPoints.getItem(2).y is 200
PASS poly.points.getItem(2).x is 200
PASS poly.points.getItem(2).y is 200
PASS successfullyParsed is true

TEST COMPLETE
